feat: implement site navigation pages and shared responsive navbar component
