+++ frontend/src/app/about/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import AboutComponent from "@/components/About";
-import { HeroHeader } from "@/components/header";
+import { HeroHeader } from "@/components/layout/navbar";
 import FooterSection from "@/components/layout/footer";
 export default function AboutPage() {
   return (
