+++ frontend/src/app/blog/page.tsx
@@ -1,5 +1,5 @@
 import BlogPageComponent from "@/components/blogpage";
-import { HeroHeader } from "@/components/header";
+import { HeroHeader } from "@/components/layout/navbar";
 import FooterSection from "@/components/layout/footer";
 
 export default function BlogPage() {
