+++ frontend/src/app/community/page.tsx
@@ -1,4 +1,4 @@
-import { HeroHeader } from "@/components/header";
+import { HeroHeader } from "@/components/layout/navbar";
 import FooterSection from "@/components/layout/footer";
 import CommunityComponent from "@/components/communitypage";
 export default function CommunityPage() {
