+++ frontend/src/app/consultations/nutritionists/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import ChooseNutritionist from "@/components/choosenutritionist";
-import { HeroHeader } from "@/components/header";
+import { HeroHeader } from "@/components/layout/navbar";
 import FooterSection from "@/components/layout/footer";
 
 export default function ConsultationPage() {
