+++ frontend/src/app/faq/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import FAQPage from "@/components/FaqPage";
-import { HeroHeader } from "@/components/header";
+import { HeroHeader } from "@/components/layout/navbar";
 import FooterSection from "@/components/layout/footer";
 export default function PrivacyPolicyPage() {
   return (
