+++ frontend/src/app/marketplace/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { HeroHeader } from "@/components/header";
+import { HeroHeader } from "@/components/layout/navbar";
 import FooterSection from "@/components/layout/footer";
 import PlanMarketplace from "@/components/planmaketplace";
