+++ frontend/src/app/privacypolicy/page.tsx
@@ -1,6 +1,6 @@
 "use clint";
 import PrivacyPolicy from "@/components/PrivacyPolicy";
-import { HeroHeader } from "@/components/header";
+import { HeroHeader } from "@/components/layout/navbar";
 import FooterSection from "@/components/layout/footer";
 export default function PrivacyPolicyPage() {
   return (
