+++ frontend/src/app/services/page.tsx
@@ -1,5 +1,5 @@
 import Services from "@/components/services";
-import { HeroHeader } from "@/components/header";
+import { HeroHeader } from "@/components/layout/navbar";
 import FooterSection from "@/components/layout/footer";
 
 export default function ServicesPage() {
