+++ frontend/src/app/subscription/page.tsx
@@ -1,5 +1,5 @@
 import SubscriptionPlans from "@/components/subscription/subscription";
-import { HeroHeader } from "@/components/header";
+import { HeroHeader } from "@/components/layout/navbar";
 import FooterSection from "@/components/layout/footer";
 
 export default function SubscriptionPage() {
