--- frontend/src/components/header.tsx
@@ -1,273 +1,0 @@
-"use client";
-
-import React, { useState, useEffect } from "react";
-import Link from "next/link";
-import { useTheme } from "next-themes";
-import { Logo } from "@/components/layout/logo";
-import { cn } from "@/lib/utils";
-import { Button } from "@/components/ui/Button";
-import { Sun, Moon } from "lucide-react";
-import {
-  Popover,
-  PopoverButton,
-  PopoverGroup,
-  PopoverPanel,
-  Dialog,
-  DialogPanel,
-  Disclosure,
-  DisclosureButton,
-  DisclosurePanel,
-} from "@headlessui/react";
-import {
-  ChevronDownIcon,
-  Bars3Icon,
-  XMarkIcon,
-  CalculatorIcon,
-  UserGroupIcon,
-  SparklesIcon,
-  CalendarIcon,
-  BookOpenIcon,
-  UsersIcon,
-  InformationCircleIcon,
-} from "@heroicons/react/24/outline";
-
-const services = [
-  {
-    name: "AI Calorie Estimation",
-    description: "Get instant nutritional insights from your food photos.",
-    href: "/services#ai-calorie-estimation",
-    icon: CalculatorIcon,
-  },
-  {
-    name: "Online Consultation",
-    description: "Connect with certified nutritionists from anywhere.",
-    href: "/services#online-consultation",
-    icon: UserGroupIcon,
-  },
-  {
-    name: "Personalized Plans",
-    description: "Custom meal and nutrition plans tailored to your goals.",
-    href: "/services#personalized-plans",
-    icon: SparklesIcon,
-  },
-  {
-    name: "Seasonal Programs",
-    description: "Specialized guides for Ramadan and Summer.",
-    href: "/services#seasonal-programs",
-    icon: CalendarIcon,
-  },
-];
-
-const mainNav = [
-  { name: "Blog", href: "/blog", icon: BookOpenIcon },
-  { name: "Community", href: "/community", icon: UsersIcon },
-  { name: "About", href: "/about", icon: InformationCircleIcon },
-];
-
-export const HeroHeader = () => {
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
-  const [isScrolled, setIsScrolled] = useState(false);
-  const { theme, setTheme, resolvedTheme } = useTheme();
-  const [mounted, setMounted] = useState(false);
-
-  useEffect(() => setMounted(true), []);
-
-  useEffect(() => {
-    const handleScroll = () => {
-      setIsScrolled(window.scrollY > 50);
-    };
-    window.addEventListener("scroll", handleScroll);
-    return () => window.removeEventListener("scroll", handleScroll);
-  }, []);
-
-  const isDark = resolvedTheme === "dark";
-  const toggleTheme = () => setTheme(isDark ? "light" : "dark");
-
-  return (
-    <header>
-      {/* OUTER NAV (CENTERING CONTAINER) */}
-      <nav className="fixed z-50 w-full flex justify-center pointer-events-none">
-        {/* ANIMATED NAVBAR SHELL */}
-        <div
-          className={cn(
-            "pointer-events-auto transition-all duration-500 ease-in-out",
-            "backdrop-blur-md",
-            "flex items-center",
-            isScrolled
-              ? "w-full lg:w-[75vw] rounded-none lg:rounded-full px-6 py-2 shadow-xl bg-white/80 dark:bg-white/10 border border-border/30 dark:border-transparent"
-              : "w-full rounded-none px-6 py-3 border border-transparent bg-transparent",
-          )}
-        >
-          {/* CONTENT WRAPPER */}
-          <div className="mx-auto w-full max-w-7xl">
-            <div className="flex items-center justify-between">
-              {/* LOGO */}
-              <div className="flex lg:flex-1">
-                <Link href="/" className="flex items-center space-x-3">
-                  <Logo />
-                </Link>
-              </div>
-
-              {/* MOBILE MENU BUTTON */}
-              <div className="flex lg:hidden">
-                <button
-                  type="button"
-                  onClick={() => setMobileMenuOpen(true)}
-                  className=" -m-2.5 inline-flex items-center justify-center
-    rounded-md p-2.5
-    text-foreground dark:text-white
-    hover:text-emerald-500 dark:hover:text-emerald-300
-    focus:outline-none"
-                >
-                  <Bars3Icon className="size-7" />
-                </button>
-              </div>
-
-              {/* DESKTOP NAV */}
-              <PopoverGroup className="hidden lg:flex lg:gap-x-8 lg:items-center">
-                <Popover className="relative">
-                  <PopoverButton
-                    className="
-    flex items-center gap-x-1
-    text-sm font-medium uppercase tracking-[0.025em]
-    text-foreground dark:text-white hover:text-emerald-500 dark:hover:text-emerald-300
-    transition-colors
-    outline-none focus:outline-none focus:ring-0
-  "
-                  >
-                    Services
-                    <ChevronDownIcon className="size-4 transition-transform group-data-[open]:rotate-180" />
-                  </PopoverButton>
-
-                  <PopoverPanel
-                    transition
-                    className="absolute -left-8 top-full z-10 mt-3 w-screen max-w-md rounded-3xl bg-popover/90 backdrop-blur-md border border-border-subtle shadow-2xl transition"
-                  >
-                    <div className="p-3">
-                      {services.map((item) => (
-                        <div
-                          key={item.name}
-                          className="flex items-center gap-x-6 rounded-2xl p-4 hover:bg-background transition"
-                        >
-                          <div className="flex size-11 items-center justify-center rounded-lg bg-secondary">
-                            <item.icon className="size-6" />
-                          </div>
-                          <div>
-                            <Link href={item.href} className="font-semibold">
-                              {item.name}
-                            </Link>
-                            <p className="text-sm text-muted-foreground">
-                              {item.description}
-                            </p>
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  </PopoverPanel>
-                </Popover>
-
-                {mainNav.map((item) => (
-                  <Link
-                    key={item.name}
-                    href={item.href}
-                    className="text-sm font-medium uppercase tracking-[0.025em] text-foreground dark:text-white hover:text-emerald-500 dark:hover:text-emerald-300"
-                  >
-                    {item.name}
-                  </Link>
-                ))}
-              </PopoverGroup>
-
-              {/* RIGHT SIDE ACTIONS */}
-              <div className="hidden lg:flex lg:flex-1 lg:justify-end lg:items-center lg:gap-x-4">
-                {/* THEME TOGGLE */}
-                <button
-                  id="theme-toggle"
-                  type="button"
-                  onClick={toggleTheme}
-                  aria-label="Toggle theme"
-                  className="relative inline-flex items-center justify-center rounded-full p-2
-                    text-foreground dark:text-white hover:text-emerald-500 dark:hover:text-emerald-300
-                    hover:bg-black/5 dark:hover:bg-white/10
-                    transition-colors duration-200
-                    focus:outline-none focus:ring-2 focus:ring-emerald-500/40"
-                >
-                  {mounted && (
-                    isDark
-                      ? <Sun className="size-5" />
-                      : <Moon className="size-5" />
-                  )}
-                </button>
-
-                {!isScrolled ? (
-                  <>
-                    <Button
-                      key="login-btn"
-                      variant="ghost"
-                      className="animate-in fade-in zoom-in duration-300"
-                      asChild
-                    >
-                      <Link
-                        href="/login"
-                        className="text-primary dark:text-white hover:text-emerald-400 dark:hover:text-emerald-300"
-                      >
-                        Login
-                      </Link>
-                    </Button>
-                    <Button
-                      key="signup-btn"
-                      className="animate-in fade-in zoom-in duration-300"
-                    >
-                      <Link href="/register">Sign Up</Link>
-                    </Button>
-                  </>
-                ) : (
-                  <Button
-                    key="get-started-btn"
-                    className="animate-in fade-in zoom-in duration-300 animate-delay-100"
-                    asChild
-                  >
-                    <Link href="/register">Get Started</Link>
-                  </Button>
-                )}
-              </div>
-            </div>
-          </div>
-        </div>
-      </nav>
-
-      {/* MOBILE DRAWER (UNCHANGED) */}
-      <Dialog
-        open={mobileMenuOpen}
-        onClose={setMobileMenuOpen}
-        className="lg:hidden"
-      >
-        <div className="fixed inset-0 z-50" />
-        <DialogPanel className="fixed inset-y-0 right-0 z-50 w-full bg-background p-6 sm:max-w-sm">
-          <div className="flex items-center justify-between">
-            <Logo />
-            <div className="flex items-center gap-2">
-              <button
-                type="button"
-                onClick={toggleTheme}
-                aria-label="Toggle theme"
-                className="inline-flex items-center justify-center rounded-full p-2
-                  text-foreground dark:text-white hover:text-emerald-500 dark:hover:text-emerald-300
-                  hover:bg-black/5 dark:hover:bg-white/10
-                  transition-colors duration-200"
-              >
-                {mounted && (
-                  isDark
-                    ? <Sun className="size-5" />
-                    : <Moon className="size-5" />
-                )}
-              </button>
-              <button onClick={() => setMobileMenuOpen(false)}>
-                <XMarkIcon className="size-7" />
-              </button>
-            </div>
-          </div>
-        </DialogPanel>
-      </Dialog>
-    </header>
-  );
-};
